requirements.txt
# Requirements for Natural Gas Trading System

# Data processing and analysis
numpy>=1.22.0
pandas>=1.4.0
scipy>=1.8.0
scikit-learn>=1.0.0
statsmodels>=0.13.0

# Data visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.6.0

# Financial analysis
vectorbt>=0.24.0
yfinance>=0.1.70
ta-lib>=0.4.24
pyfolio>=0.9.2

# API access
requests>=2.27.0
python-dotenv>=0.19.0
pyyaml>=6.0.0

# Web scraping
beautifulsoup4>=4.10.0
lxml>=4.7.0

# Data storage
pyarrow>=7.0.0
sqlalchemy>=1.4.0
pymongo>=4.0.0  # Optional for NoSQL storage

# API clients
eia-python>=1.22
noaa-sdk>=0.1.2

# Web dashboard
streamlit>=1.8.0
dash>=2.3.0  # Optional for more complex dashboards

# Scheduling and automation
schedule>=1.1.0
apscheduler>=3.9.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0
hypothesis>=6.0.0

# Development tools
black>=22.1.0
flake8>=4.0.0
isort>=5.10.0
pre-commit>=2.17.0
ipykernel>=6.9.0

# Machine learning
tensorflow>=2.8.0  # Optional for deep learning
torch>=1.11.0  # Optional for deep learning
xgboost>=1.5.0
lightgbm>=3.3.0
sklearn-pandas>=2.2.0

# Time series specific
prophet>=1.0.1  # Optional for time series forecasting

# Logging and monitoring
rich>=12.0.0  # Better console output
prometheus_client>=0.13.0  # Optional for metrics

# Documentation
sphinx>=4.4.0  # Optional for auto-documentation
mkdocs>=1.3.0  # Optional for documentation

# Deployment
docker>=5.0.0  # Optional for containerization
fastapi>=0.75.0  # Optional for API services

# Jupyter notebooks
jupyter>=1.0.0

# Airflow (optional, commented out for initial setup)
# apache-airflow>=2.2.0 